+++ src/lib/user.ts
@@ -1,0 +1,10 @@
+export type User = {
+    id: number,
+    name: string,
+    email: string,
+    picture: string,
+    locale: string,
+    created_at: string,
+    updated_at: string,
+    access_token: string,
+}
